Revert "RN-333 allow account deletion in medtkapp (#3904)"
This reverts commit 477479deb67ab8cf10a990e92defc20ba3603974.
