+++ packages/central-server/.env.example
@@ -4,7 +4,7 @@ AWS_ACCESS_KEY_ID=
 AWS_SECRET_ACCESS_KEY=
 CLIENT_SECRET=
 CLIENT_USERNAME=
-TUPAIA_ADMIN_EMAIL_ADDRESS=
+COUNTRY_REQUEST_EMAIL_ADDRESS=
 DB_ENABLE_SSL=
 DB_NAME=
 DB_PASSWORD=
