--- packages/central-server/src/apiV2/utilities/getUserInfoInString.js
@@ -1,10 +1,0 @@
-import { DatabaseError } from '@tupaia/utils';
-
-export const getUserInfoInString = async (userId, models) => {
-  try {
-    const user = await models.user.findById(userId);
-    return `${user.first_name} ${user.last_name} (${user.email} - ${user.id}), ${user.position} at ${user.employer}, `;
-  } catch (error) {
-    throw new DatabaseError('finding user', error);
-  }
-};
