+++ packages/central-server/src/apiV2/utilities/index.js
@@ -40,4 +40,3 @@ export {
   translateUserEmailToIdAndAssessorName,
 } from './translateSurveyResponseFields';
 export { translateObjectFields } from './translateObjectFields';
-export { getUserInfoInString } from './getUserInfoInString';
