+++ packages/meditrak-app/app/navigation/constants.js
@@ -13,7 +13,6 @@ export const SYNC_SCREEN = 'SyncScreen';
 export const SURVEY_SCREEN = 'SurveyScreen';
 export const SURVEYS_MENU_SCREEN = 'SurveysMenu';
 export const CHANGE_PASSWORD_SCREEN = 'ChangePassword';
-export const DELETE_ACCOUNT_REQUEST_SCREEN = 'DeleteAccount';
 export const WEB_BROWSER_SCREEN = 'WEB_BROWSER_SCREEN';
 export const REALM_EXPLORER_SCREEN = 'RealmExplorer';
 export const HOME_SCREEN = 'HOME_SCREEN';
