+++ packages/meditrak-app/app/reducers.js
@@ -16,7 +16,6 @@ import { reducer as userLocation } from './utilities/userLocation';
 import { reducer as rewards } from './rewards';
 import { reducer as messages } from './messages';
 import { reducer as changePassword } from './changePassword';
-import { reducer as requestAccountDeletion } from './requestAccountDeletion';
 
 export const reducers = {
   authentication,
@@ -33,5 +32,4 @@ export const reducers = {
   rewards,
   messages,
   changePassword,
-  requestAccountDeletion,
 };
