--- packages/meditrak-app/app/requestAccountDeletion/index.js
@@ -1,7 +1,0 @@
-/**
- * Tupaia MediTrak
- * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
- */
-
-export { reducer } from './reducer';
-export { RequestAccountDeletionContainer } from './RequestAccountDeletionContainer';
