+++ packages/meditrak-app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tupaia/meditrak-app",
-  "version": "1.11.122",
+  "version": "1.11.121",
   "private": true,
   "description": "Android and iOS app for surveying medical facilities",
   "repository": {
